+++ partials/post-layouts/archive-post.php
@@ -24,7 +24,7 @@ $youtube_id = $show_video_embed && ! empty( $meta['_cmb_utube'][0] ) ? $meta['_c
   <?php } elseif ( $is_show_tags ) { ?>
     <div class="layout-thumbnail-frame">
       <div class="layout-thumbnail-frame__inner mt-1 ml-1">
-        <?php render_post_ui_tags( $this_post_id, true, true ); ?>
+        <?php render_post_ui_tags( $this_post_id, true, true, 'no-border' ); ?>
       </div>
       <a href="<?php the_permalink(); ?>" class="ui-hover u-display-block">
         <?php
@@ -57,8 +57,6 @@ switch ( $text_size ) {
   case 'regular':
     ?>
     <h5 class="index-post-title font-size-9 font-weight-bold mt-2 text-wrap-pretty"><?php the_title(); ?></h5>
-    <?php
-    ?>
     <h6 class="font-size-8 font-weight-bold text-uppercase mt-1 text-wrap-pretty">
       <?php
       if ( $is_article ) {
@@ -82,9 +80,7 @@ switch ( $text_size ) {
   case 'large':
     ?>
     <h3 class="font-size-10 font-weight-bold mt-2 text-wrap-pretty"><?php the_title(); ?></h3>
-    <?php
-    ?>
-  <h3 class="font-size-9 font-weight-bold text-wrap-pretty">
+    <h3 class="font-size-9 font-weight-bold text-wrap-pretty">
       <?php
       if ( $is_article ) {
         render_bylines( $this_post_id );