+++ category.php
@@ -153,9 +153,10 @@ if( have_posts() ) {
   while( have_posts() ) {
     the_post();
 
-    get_template_part('partials/post-layouts/flex-post', null, array(
+    get_template_part('partials/post-layouts/archive-post', null, array(
       'grid-item-classes' => 'grid-item is-s-24 is-l-12 is-xxl-8 mb-4',
       'image-size' => 'col12-16to9',
+      'show-tags' => true,
     ));
   }
 } else {